user_service/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
mysqlclient
PyJWT
python-dotenv
requests
